requirements.txt
allure-pytest==2.11.1
pytest==7.1.2
requests==2.28.1
selenium==4.4.3

validators~=0.18.2
Faker==15.1.0